# [pytorch][cuda] Generate kernels for 5x5 filters on depth wise convolution backward