@eqy 2024-06-26
Am I reading the benchmark results correctly in that the improvement seems to be mainly for fp32?